Moved classes to their right packages
